Reworking build system, framework
Instead of trying to bundle the raw libs into frameworks...
* encapsulate the libs into 'OgvKit' static library project
* expose only the Obj-C wrapper interface

+++ OgvKit/OgvKit/OGVDecoder.m
@@ -200,7 +200,7 @@
 			
             self.hasAudio = YES;
             self.audioChannels = vi.channels;
-            self.audioRate = vi.rate;
+            self.audioRate = (int)vi.rate;
 		}
         
         appState = STATE_DECODING;
